convert to custom resource
