+++ Gemfile
@@ -10,3 +10,5 @@ gem 'tomlrb'
 gem 'rake'
 gem 'stove'
 gem 'community_cookbook_releaser'
+gem 'chefspec'
+gem 'berkshelf'
